+++ src/actions/index.js
@@ -1,5 +1,3 @@
 export const CREATE_EVENT = "CREATE_EVENT"
 export const DELETE_EVENT = "DELETE_EVENT"
-
 export const COMPLETE_EVENT = "COMPLETE_EVENT"
-export const DELETE_COMPLETE_EVENT = "DELETE_COMPLETE_EVENT"
